better check daemon

+++ penmode.py
@@ -62,17 +62,17 @@ class MainWindow ( QMainWindow , Ui_MainWindow):
 			self.ui.skipfishTarget.setText(self.pencore.get_target())
 			self.ui.sqlmapTarget.setText(self.pencore.get_target())
 		
-		if self.checkSocat() == 1:
-			self.ui.pushSocat.setText('Enabled')
-			
 		if self.checkTor() == 1:
-			self.ui.pushTor.setText('Enabled')
+			self.ui.pushTor.setText('Disable')
+		else:
+			self.startStopTor()
 		
 		signal.signal(signal.SIGINT, signal.SIG_DFL)
 		self.show()
 	
 	def showTerminal(self,cmd,area):
-		area.append(cmd)
+		self.startStopSocat()
+		self.startStopTor()
 		p = Popen(cmd + ' | sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[m|K]//g"', shell=True, stdout=PIPE).stdout
 		stdout = ''
 		while True:
@@ -87,23 +87,24 @@ class MainWindow ( QMainWindow , Ui_MainWindow):
 		return self.pencore.check_socat()
 	
 	def startStopSocat(self):
-		if self.checkSocat() == 1:
-			Popen('killall socat', shell=True, stdout=PIPE)
-			self.ui.pushSocat.setText('Enable')
-		else:
-			Popen('socat TCP4-LISTEN:8080,fork SOCKS4a:127.0.0.1:$ip,socksport=9050 &', shell=True, stdout=PIPE)
-			self.ui.pushSocat.setText('Disable')
+		if self.pencore.get_target() != None:
+			if self.checkSocat() == 1:
+				Popen('killall socat', shell=True, stdout=PIPE)
+				self.ui.pushSocat.setText('Enable')
+			else:
+				self.pencore.start_socat()
+				self.ui.pushSocat.setText('Disable')
 			
 	def checkTor(self):
 		return self.pencore.check_tor()
 	
 	def startStopTor(self):
 		if self.checkTor() == 1:
-			Popen('su-to-root -X -c /etc/init.d/tor stop', shell=True, stdout=PIPE)
-			self.ui.pushSocat.setText('Enable')
+			Popen('/etc/init.d/tor stop', shell=True, stdout=PIPE)
+			self.ui.pushTor.setText('Enable')
 		else:
-			Popen('su-to-root -X -c /etc/init.d/tor start', shell=True, stdout=PIPE)
-			self.ui.pushSocat.setText('Disable')
+			self.pencore.start_tor()
+			self.ui.pushTor.setText('Disable')
 			
 	
 	def checkTarget(self,target):
